requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.8.2
pyyaml==6.0.2
prometheus-client==0.20.0
azure-identity==1.17.1
msal==1.28.0
requests==2.32.3
python-dotenv==1.0.1
aiofiles==23.2.1
PyJWT[crypto]==2.9.0
tenacity==8.2.3
jinja2==3.1.4
aiohttp==3.9.5
schedule==1.2.0
